msm: kgsl: Add in-stream support for IOMMU-v1
Add support to perform in-stream pagetable updates and tlb
flushes for IOMMU-v1. This is only supported if the IOMMU
hardware includes halt lock supprt. The IOMMU halt lock is
held when writing to TTBR0 register of IOMMU unit from the GPU.
This prevents the GPU pre-fetcher from using the IOMMU
while TTBR0 is being updated.

Change-Id: I43dab204d9f10a97fc1db53d120dcab1f5e9a5af
Signed-off-by: Shubhraprakash Das <sadas@codeaurora.org>
Signed-off-by: Jeremy Gebben <jgebben@codeaurora.org>
Signed-off-by: Carter Cooper <ccooper@codeaurora.org>

Conflicts:
	drivers/gpu/msm/adreno.c
	drivers/gpu/msm/adreno.h
	drivers/gpu/msm/kgsl_iommu.c
	drivers/gpu/msm/kgsl_mmu.h